requirements.txt
anyio==3.7.0
click==8.1.3
dnspython==2.3.0
email-validate==1.1.2
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
h11==0.14.0
httptools==0.5.0
idna==2.10
jose==1.0.0
passlib==1.7.4
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
